requirements.txt
pandas
numpy
python-dotenv
PySide6